fix: incremental refreshers first execution parses only recently updated data
FullRefresher is executed first, and its purpose is to cache all data.
Incremental refresher is executed every 5 seconds, its purpose is to cache recently updated data.

On first execution, lastRefreshAt was set to -1, so the incremental refresher was loading all the data.
Ths fix set the lastRefreshAt to a relevant time, before start of service.
This ensure to load only recent data, without any loss between full and incremental refresh.